requirements.txt
pandas
numpy
matplotlib
google-cloud-bigquery
google-cloud-bigquery[pandas]
google-cloud-storage
numba
polars
dask
distributed

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=2.0.0
sphinx-autodoc-typehints>=1.25.2
sphinx-copybutton>=0.5.2
sphinx-toolbox>=3.5.0